🚚💥 Refactor: move files to src, eslint files, config exemple
