--- commands/other/test.js
@@ -1,12 +1,0 @@
-const { Message, Client } = require("discord.js");
-
-module.exports = {
-    name: "test",
-    aliases: ["t"],
-    permissions: ["SEND_MESSAGES"],
-    description: "",
-
-    run: async (client, message, args) => {
-        message.channel.send("test");
-    },
-};
